fix chat

+++ lib/app/modules/destination_details/controllers/destination_details_controller.dart
@@ -62,7 +62,7 @@ class DestinationDetailsController extends GetxController {
       title: const Text('تقييم الوجهة'),
       submitButtonText: 'تقييم',
       onSubmitted: (response) {
-        rating.value = response.rating;
+        rating.value = response.rating.toInt();
         doRateDestination(guestId, destinationId);
       },
     ));
